ref(settings-page): adopt coherent naming convention
