+++ src/app/(user)/cdp/[study]/settings/clients/clients.tsx
@@ -4,12 +4,9 @@ import { useState } from 'react';
 
 import { Box, BoxContent, BoxHeader, BoxTitle } from '@/components/boxes/boxes';
 import { AddButton } from '@/components/buttons';
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog';
-import { reloadWindow } from '@/lib/utils';
 
-import { addClient } from './action';
-import { ClientEditor } from './edit-client';
-import { ClientAdder } from './new-client';
+import { DisplayStudyClient } from './display-client';
+import { NewStudyClientDialog } from './new-client';
 import { StudyClientsFormType } from './schema';
 
 interface StudyInfosParamsEditorParams extends StudyClientsFormType {
@@ -32,23 +29,16 @@ export function StudyClientsParamsEditor({
                 </BoxHeader>
                 <BoxContent>
                     {clients.map((client, i) => (
-                        <ClientEditor client={client} key={i} />
+                        <DisplayStudyClient client={client} key={i} />
                     ))}
                     <AddButton text="Ajouter un client" onClick={() => setClientAddOpen(true)} />
                 </BoxContent>
             </Box>
-            <Dialog open={clientAddOpen} onOpenChange={() => setClientAddOpen(false)}>
-                <DialogContent>
-                    <DialogHeader>
-                        <DialogTitle>Nouveau client</DialogTitle>
-                        <ClientAdder
-                            onSubmit={(client) =>
-                                addClient(studyId, client).then(() => reloadWindow())
-                            }
-                        />
-                    </DialogHeader>
-                </DialogContent>
-            </Dialog>
+            <NewStudyClientDialog
+                clientAddOpen={clientAddOpen}
+                setClientAddOpen={setClientAddOpen}
+                studyId={studyId}
+            />
         </>
     );
 }
