+++ src/app/(user)/cdp/[study]/settings/clients/edit-client.tsx
@@ -1,73 +1,37 @@
 import { zodResolver } from '@hookform/resolvers/zod';
-import { CompanySize } from '@prisma/client';
-import { Dispatch, SetStateAction, useState } from 'react';
-import { FormProvider, useForm } from 'react-hook-form';
-import { FaEye, FaPencil, FaTrash } from 'react-icons/fa6';
-
-import { IconButton } from '@/components/buttons';
-import { LoadingFullStops } from '@/components/loading';
-import { DropdownManyFormElement } from '@/components/meta-components/form/dropdownMany';
-import { DropdownSingleFormElement } from '@/components/meta-components/form/dropdownSingle';
-import { InputFormElement } from '@/components/meta-components/form/input';
 import {
     AlertDialog,
     AlertDialogCancel,
     AlertDialogContent,
-    AlertDialogDescription,
-    AlertDialogFooter,
-    AlertDialogHeader,
     AlertDialogTitle,
-} from '@/components/ui/alert-dialog';
+} from '@radix-ui/react-alert-dialog';
+import { Checkbox } from '@radix-ui/react-checkbox';
+import { Dispatch, SetStateAction, useState } from 'react';
+import { FormProvider, useForm } from 'react-hook-form';
+
+import { LoadingFullStops } from '@/components/loading';
+import { DropdownManyFormElement } from '@/components/meta-components/form/dropdownMany';
+import { DropdownSingleFormElement } from '@/components/meta-components/form/dropdownSingle';
+import { InputFormElement } from '@/components/meta-components/form/input';
+import { AlertDialogFooter, AlertDialogHeader } from '@/components/ui/alert-dialog';
 import { Button } from '@/components/ui/button';
-import { Checkbox } from '@/components/ui/checkbox';
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog';
 import { Separator } from '@/components/ui/separator';
 import { COMPANY_SIZE_NAMES, COMPANY_SIZES, DOMAIN_NAMES, DOMAINS } from '@/db/types';
-import { personName, reloadWindow } from '@/lib/utils';
+import { dbg, personName, reloadWindow } from '@/lib/utils';
 
-import { removeClient, updateClient } from './action';
+import { editStudyClient } from './actions/edit-client';
 import { clientFormSchema, ClientFormType } from './schema';
 
-interface EditClientDialogProps {
-    isOpen: boolean;
-    setIsOpen: Dispatch<SetStateAction<boolean>>;
-    client: ClientFormType;
+interface EnableBoxProps {
+    label: string;
+    onCheckedChange: (value: boolean) => void;
+    checked: boolean;
 }
 
-function EditClientDialog({ isOpen, setIsOpen, client }: EditClientDialogProps) {
-    const clientId = client.clientId;
-    const [isLoading, setIsLoading] = useState(false);
-
-    return (
-        <AlertDialog open={isOpen} onOpenChange={setIsOpen}>
-            <AlertDialogContent>
-                <AlertDialogHeader>
-                    <AlertDialogTitle>Edition de {personName(client)}</AlertDialogTitle>
-                </AlertDialogHeader>
-                {clientId === undefined ? (
-                    <p>Erreur lors du chargement du client</p>
-                ) : (
-                    <EditClientForm
-                        client={client}
-                        onSubmit={(newClientData) => {
-                            setIsLoading(true);
-                            updateClient(clientId, newClientData).then(() => reloadWindow());
-                        }}
-                        isLoading={isLoading}
-                    />
-                )}
-                <AlertDialogFooter>
-                    <AlertDialogCancel className="w-full">Annuler</AlertDialogCancel>
-                </AlertDialogFooter>
-            </AlertDialogContent>
-        </AlertDialog>
-    );
-}
-
-function EnableBox({ label, onClick }: { label: string; onClick: (value: boolean) => void }) {
+function EnableBox({ label, onCheckedChange, checked }: EnableBoxProps) {
     return (
         <div className="flex items-center gap-x-2">
-            <Checkbox onCheckedChange={onClick} />
+            <Checkbox onCheckedChange={onCheckedChange} checked={checked} />
             <p>{label}</p>
         </div>
     );
@@ -80,26 +44,16 @@ interface EditClientFormProps {
 }
 
 function EditClientForm({ onSubmit, client, isLoading }: EditClientFormProps) {
-    const defaultValues = client;
-
-    if (defaultValues.company === undefined) {
-        defaultValues.company = {
-            name: '',
-            address: undefined,
-            size: CompanySize.MicroEntreprise,
-            ca: 0,
-            domains: [],
-        };
-    }
-
     const form = useForm<ClientFormType>({
         resolver: zodResolver(clientFormSchema),
         defaultValues: client,
     });
 
-    const [editCompany, setEditCompany] = useState(false);
-    const [editAddress, setEditAddress] = useState(false);
-    const [editCompanyAddress, setEditCompanyAddress] = useState(false);
+    dbg(client, '=============== editing client ==============');
+
+    const [editCompany, setEditCompany] = useState(!!client.address?.id);
+    const [editAddress, setEditAddress] = useState(!!client.company?.id);
+    const [editCompanyAddress, setEditCompanyAddress] = useState(!!client.company?.address?.id);
 
     return (
         <>
@@ -118,7 +72,11 @@ function EditClientForm({ onSubmit, client, isLoading }: EditClientFormProps) {
                     <InputFormElement form={form} label="Numéro de téléphone" name="number" />
                     <Separator />
                     <h2>Addresse du client</h2>
-                    <EnableBox onClick={setEditAddress} label="Addresse du client" />
+                    <EnableBox
+                        checked={editAddress}
+                        onCheckedChange={setEditAddress}
+                        label="Addresse du client"
+                    />
                     {editAddress && (
                         <>
                             <div className="grid-cols-5 grid gap-x-main">
@@ -154,7 +112,11 @@ function EditClientForm({ onSubmit, client, isLoading }: EditClientFormProps) {
                     )}
                     <Separator />
                     <h2>L'entreprise</h2>
-                    <EnableBox onClick={setEditCompany} label="Entreprise du client" />
+                    <EnableBox
+                        checked={editCompany}
+                        onCheckedChange={setEditCompany}
+                        label="Entreprise du client"
+                    />
                     {editCompany && (
                         <>
                             <InputFormElement form={form} label="Nom" name="company.name" />
@@ -181,7 +143,8 @@ function EditClientForm({ onSubmit, client, isLoading }: EditClientFormProps) {
                             <Separator />
                             <h2>Adresse de 'entreprise</h2>
                             <EnableBox
-                                onClick={setEditCompanyAddress}
+                                checked={editCompanyAddress}
+                                onCheckedChange={setEditCompanyAddress}
                                 label="Adresse de l'entreprise"
                             />
                             {editCompanyAddress && (
@@ -251,169 +214,38 @@ function EditClientForm({ onSubmit, client, isLoading }: EditClientFormProps) {
     );
 }
 
-interface ViewClientDialogProps {
+interface EditClientDialogProps {
     isOpen: boolean;
     setIsOpen: Dispatch<SetStateAction<boolean>>;
     client: ClientFormType;
 }
 
-function ViewClientDialog({ isOpen, setIsOpen, client }: ViewClientDialogProps) {
-    return (
-        <Dialog open={isOpen} onOpenChange={() => setIsOpen((isOpen) => !isOpen)}>
-            <DialogContent>
-                <DialogHeader>
-                    <DialogTitle>{personName(client)}</DialogTitle>
-                    <div className="grid grid-cols-2">
-                        <p>Métier</p>
-                        <p>{client.job}</p>
-                        <p>Email</p>
-                        <p>{client.email || 'Inconnu'}</p>
-                        <p>Numéro de téléphone</p>
-                        <p>{client.number || 'Inconnu'}</p>
-                        <p>Adresse</p>
-                        <p>
-                            {client.address === undefined
-                                ? 'Inconnue'
-                                : `${client.address.number} ${client.address.street}, ${client.address.zipCode}, ${client.address.city}, ${client.address.country}`}
-                            ,
-                        </p>
-                    </div>
-                    {client.company === undefined ? (
-                        <p>Entreprise non spécifiée</p>
-                    ) : (
-                        <>
-                            <Separator />
-                            <h2>Son entreprise</h2>
-                            <div className="grid grid-cols-2">
-                                <p>Nom</p>
-                                <p>{client.company.name}</p>
-                                <p>Taille</p>
-                                <p>
-                                    {client.company.size === undefined
-                                        ? 'Inconnu'
-                                        : COMPANY_SIZES[client.company.size].display}
-                                </p>
-                                <p>CA</p>
-                                <p>{client.company.ca || 'Inconnu'}</p>
-                                <p>Domaines</p>
-                                <div>
-                                    {client.company.domains.map((domain, i) => (
-                                        <p key={i}>{DOMAINS[domain].display}</p>
-                                    ))}
-                                </div>
-                                <p>Adresse</p>
-                                <p>
-                                    {client.company.address === undefined
-                                        ? 'Inconnue'
-                                        : `${client.company.address.number} ${client.company.address.street}, ${client.company.address.zipCode}, ${client.company.address.city}, ${client.company.address.country}`}
-                                    ,
-                                </p>
-                            </div>
-                        </>
-                    )}
-                </DialogHeader>
-            </DialogContent>
-        </Dialog>
-    );
-}
-
-interface DeleteClientDialogProps {
-    isOpen: boolean;
-    setIsOpen: Dispatch<SetStateAction<boolean>>;
-    studyClientId: string;
-}
-
-function DeleteClientDialog({ isOpen, setIsOpen, studyClientId }: DeleteClientDialogProps) {
+export function EditClientDialog({ isOpen, setIsOpen, client }: EditClientDialogProps) {
+    const clientId = client.clientId;
     const [isLoading, setIsLoading] = useState(false);
 
     return (
-        <AlertDialog open={isOpen} onOpenChange={() => setIsOpen((isOpen) => !isOpen)}>
+        <AlertDialog open={isOpen} onOpenChange={setIsOpen}>
             <AlertDialogContent>
                 <AlertDialogHeader>
-                    <AlertDialogTitle>Confirmer la suppression du client</AlertDialogTitle>
-                    <AlertDialogDescription>
-                        Ce client sera enlevé de cette étude. Il existera toujours dans la base si
-                        vous désirez l'ajouter plus tard. Changer le client risque de corrompre
-                        certaines données si l'étude est trop avancée.
-                    </AlertDialogDescription>
-                    {isLoading ? (
-                        <div className="w-full items-center flex justify-center">
-                            <LoadingFullStops />
-                        </div>
-                    ) : (
-                        <div className="w-full items-center flex justify-between pt-4">
-                            <Button
-                                variant="outline"
-                                onClick={() => {
-                                    setIsOpen(false);
-                                }}
-                            >
-                                Annuler la suppression
-                            </Button>
-                            <Button
-                                variant="default"
-                                onClick={() => {
-                                    setIsLoading(true);
-                                    removeClient(studyClientId).then(() => reloadWindow());
-                                }}
-                            >
-                                Confirmer la suppression
-                            </Button>
-                        </div>
-                    )}
+                    <AlertDialogTitle>Edition de {personName(client)}</AlertDialogTitle>
                 </AlertDialogHeader>
+                {clientId === undefined ? (
+                    <p>Erreur lors du chargement du client</p>
+                ) : (
+                    <EditClientForm
+                        client={client}
+                        onSubmit={(newClientData) => {
+                            setIsLoading(true);
+                            editStudyClient(clientId, newClientData).then(() => reloadWindow());
+                        }}
+                        isLoading={isLoading}
+                    />
+                )}
+                <AlertDialogFooter>
+                    <AlertDialogCancel className="w-full">Annuler</AlertDialogCancel>
+                </AlertDialogFooter>
             </AlertDialogContent>
         </AlertDialog>
-    );
-}
-
-export function ClientEditor({ client }: { client: ClientFormType }) {
-    const [isDeleteOpen, setIsDeleteOpen] = useState(false);
-    const [isEditorOpen, setIsEditorOpen] = useState(false);
-    const [isViewOpen, setIsViewOpen] = useState(false);
-    const studyClientId: string | undefined = client.studyClientId;
-
-    return (
-        <div className="flex justify-between items-center border rounded-md pl-2 border-input">
-            {studyClientId === undefined ? (
-                <p>Failed to load client</p>
-            ) : (
-                <>
-                    <p>{personName(client)}</p>
-                    <div>
-                        <IconButton
-                            hoverContent="See client information"
-                            Icon={FaEye}
-                            onClick={() => setIsViewOpen(true)}
-                        />
-                        <IconButton
-                            hoverContent="Update client infromation"
-                            onClick={() => setIsEditorOpen(true)}
-                            Icon={FaPencil}
-                        />
-                        <IconButton
-                            hoverContent="Remove client from study"
-                            Icon={FaTrash}
-                            onClick={() => setIsDeleteOpen(true)}
-                        />
-                    </div>
-                    <DeleteClientDialog
-                        isOpen={isDeleteOpen}
-                        setIsOpen={setIsDeleteOpen}
-                        studyClientId={studyClientId}
-                    />
-                    <EditClientDialog
-                        isOpen={isEditorOpen}
-                        setIsOpen={setIsEditorOpen}
-                        client={client}
-                    />
-                    <ViewClientDialog
-                        isOpen={isViewOpen}
-                        setIsOpen={setIsViewOpen}
-                        client={client}
-                    />
-                </>
-            )}
-        </div>
     );
 }
