+++ src/app/(user)/cdp/[study]/settings/clients/new-client.tsx
@@ -1,12 +1,16 @@
 import { zodResolver } from '@hookform/resolvers/zod';
+import { Dispatch, SetStateAction } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 
 import { InputFormElement } from '@/components/meta-components/form/input';
 import { Button } from '@/components/ui/button';
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog';
+import { reloadWindow } from '@/lib/utils';
 
+import { newStudyClient } from './actions';
 import { clientFormSchema, ClientFormType } from './schema';
 
-export function ClientAdder({ onSubmit }: { onSubmit: (values: ClientFormType) => void }) {
+function NewStudyClientForm({ onSubmit }: { onSubmit: (values: ClientFormType) => void }) {
     const form = useForm<ClientFormType>({
         resolver: zodResolver(clientFormSchema),
     });
@@ -27,36 +31,6 @@ export function ClientAdder({ onSubmit }: { onSubmit: (values: ClientFormType) =
                     />
                     <InputFormElement form={form} label="Prénom" name="firstName" />
                     <InputFormElement form={form} label="Nom de famille" name="lastName" />
-                    <InputFormElement form={form} label="Numéro de téléphone" name="number" />
-                    <div className="grid-cols-5 grid gap-x-main">
-                        <InputFormElement
-                            form={form}
-                            className="mb-0"
-                            label="N°"
-                            name="address.number"
-                        />
-                        <InputFormElement
-                            className="col-span-4 mb-0"
-                            form={form}
-                            label="Rue"
-                            name="address.street"
-                        />
-                    </div>
-                    <div className="grid-cols-2 grid gap-x-main">
-                        <InputFormElement
-                            className="mb-0"
-                            form={form}
-                            label="Ville"
-                            name="address.city"
-                        />
-                        <InputFormElement
-                            className="mb-0"
-                            form={form}
-                            label="Code postal"
-                            name="address.zipCode"
-                        />
-                    </div>
-                    <InputFormElement form={form} label="Pays" name="address.country" />
                 </form>
             </FormProvider>
             <Button
@@ -70,5 +44,30 @@ export function ClientAdder({ onSubmit }: { onSubmit: (values: ClientFormType) =
                 Valider
             </Button>
         </>
+    );
+}
+
+export function NewStudyClientDialog({
+    studyId,
+    clientAddOpen,
+    setClientAddOpen,
+}: {
+    studyId: string;
+    clientAddOpen: boolean;
+    setClientAddOpen: Dispatch<SetStateAction<boolean>>;
+}) {
+    return (
+        <Dialog open={clientAddOpen} onOpenChange={setClientAddOpen}>
+            <DialogContent>
+                <DialogHeader>
+                    <DialogTitle>Nouveau client</DialogTitle>
+                    <NewStudyClientForm
+                        onSubmit={(client) =>
+                            newStudyClient(studyId, client).then(() => reloadWindow())
+                        }
+                    />
+                </DialogHeader>
+            </DialogContent>
+        </Dialog>
     );
 }
